+++ checkov/yaml_doc/base_registry.py
@@ -41,6 +41,7 @@ class Registry(BaseCheckRegistry):
                             results,
                             scanned_file,
                             skip_info,
+                            entity
                         )
             if isinstance(analyzed_entities, list):
                 for item in analyzed_entities:
@@ -53,6 +54,7 @@ class Registry(BaseCheckRegistry):
                             results,
                             scanned_file,
                             skip_info,
+                            entity
                         )
         if isinstance(entity, list):
             for item in entity:
@@ -65,6 +67,7 @@ class Registry(BaseCheckRegistry):
                         results,
                         scanned_file,
                         skip_info,
+                        entity
                     )
                     if result == CheckResult.FAILED:
                         break
@@ -91,7 +94,7 @@ class Registry(BaseCheckRegistry):
             entity_type: str, results: Dict[str, Any]
     ) -> None:
         self.update_result(
-            check, entity, entity_name, entity_type, results, scanned_file, skip_info
+            check, entity, entity_name, entity_type, results, scanned_file, skip_info, entity
         )
 
     def _scan_yaml(
@@ -174,12 +177,13 @@ class Registry(BaseCheckRegistry):
     def update_result(
             self,
             check: BaseCheck,
-            entity_configuration: Dict[str, Any],
+            entity_configuration: dict[str, Any],
             entity_name: str,
             entity_type: str,
             results: Dict[str, Any],
             scanned_file: str,
             skip_info: _SkippedCheck,
+            definition: dict[str, Any]
     ) -> CheckResult:
         check_result = self.run_check(
             check,
@@ -194,7 +198,8 @@ class Registry(BaseCheckRegistry):
                                          entity_name,
                                          entity_type,
                                          scanned_file,
-                                         skip_info)
+                                         skip_info,
+                                         definition)
 
         result = check_result["result"]
 
@@ -222,13 +227,16 @@ class Registry(BaseCheckRegistry):
         return result
 
     def get_result_key(self, check: BaseCheck,
-                       entity_configuration: Dict[str, Any],
+                       entity_configuration: dict[str, Any],
                        entity_name: str,
                        entity_type: str,
                        scanned_file: str,
-                       skip_info: _SkippedCheck) -> str:
+                       skip_info: _SkippedCheck,
+                       definition: dict[str, Any]) -> str:
         if STARTLINE_MARK in entity_configuration and ENDLINE_MARK in entity_configuration:
-            return f'{entity_type}.{entity_name}.{check.id}[{entity_configuration[STARTLINE_MARK]}:{entity_configuration[ENDLINE_MARK]}]'
+            key = f'{entity_type}.{entity_name}.{check.id}[{entity_configuration[STARTLINE_MARK]}:{entity_configuration[ENDLINE_MARK]}]'
+            return Registry.modify_gha_key(key, check, definition)
+
         if isinstance(entity_configuration, list):
             start_line = None
             end_line = None
@@ -245,9 +253,43 @@ class Registry(BaseCheckRegistry):
                     if subconf_startline < start_line:
                         start_line = subconf_startline
             if start_line and end_line:
-                return f'{entity_type}.{entity_name}.{check.id}[{start_line}:{end_line}]'
-        return f'{entity_type}.{entity_name}.{check.id}'
+                key = f'{entity_type}.{entity_name}.{check.id}[{start_line}:{end_line}]'
+                return Registry.modify_gha_key(key, check, definition)
+
+        key = f'{entity_type}.{entity_name}.{check.id}'
+        return Registry.modify_gha_key(key, check, definition)
 
     def extract_entity_details(self, entity: dict[str, Any]) -> tuple[str, str, dict[str, Any]]:
         # not used, but is an abstractmethod
         pass
+
+    @staticmethod
+    def modify_gha_key(key: str, check: BaseCheck, definition: dict[str, Any]) -> str:
+        if 'GITHUB_ACTION' in check.bc_id:
+            potential_job_name = key.split('.')[1]
+            if potential_job_name != '*':
+                new_key = f'jobs.{potential_job_name}'
+            else:
+                start_line, end_line = Registry.get_start_and_end_lines(key)
+                job_name = Registry.resolve_job_name(definition, int(start_line), int(end_line))
+                new_key = f'jobs.{job_name}.steps'
+            return new_key
+        return key
+
+    @staticmethod
+    def get_start_and_end_lines(key: str) -> list[str]:
+        check_name = key.split('.')[-1]
+        try:
+            start_end_line_bracket_index = check_name.index('[')
+        except ValueError:
+            return ['-1', '-1']
+        return check_name[start_end_line_bracket_index + 1: len(check_name) - 1].split(':')
+
+    @staticmethod
+    def resolve_job_name(definition: dict[str, Any], start_line: int, end_line: int) -> str:
+        for key, job in definition.get('jobs', {}).items():
+            if key in [STARTLINE_MARK, ENDLINE_MARK]:
+                continue
+            if job[STARTLINE_MARK] <= start_line <= end_line <= job[ENDLINE_MARK]:
+                return str(key)
+        return ""
