+++ tests/generic_yaml/test_registry.py
@@ -1,0 +1,47 @@
+import pytest
+
+from checkov.github_actions.checks.job.SuspectCurlInScript import SuspectCurlInScript
+from checkov.yaml_doc.base_registry import Registry
+
+
+def test_get_start_and_end_lines():
+    keys = ['jobs.*.steps[].jobs.*.steps[].CKV_GHA_3[18:22]',
+            'jobs.*.steps[].CKV_GHA_3[18:22]',
+            'jobs.job_name.CKV_GHA_3[18:22]']
+
+    for key in keys:
+        start_line, end_line = Registry.get_start_and_end_lines(key)
+        assert start_line == '18'
+        assert end_line == '22'
+
+
+@pytest.mark.parametrize(
+    "start_line,end_line,expected_job_name",
+    [
+        (10, 15, "container-test-job"),
+        (8, 20, "container-test-job"),
+        (24, 30, "second_job"),
+        (25, 27, "second_job"),
+        (5, 40, "")
+    ],
+)
+def test_resolve_job_name(start_line, end_line, expected_job_name, definition):
+    job_name = Registry.resolve_job_name(definition, start_line, end_line)
+
+    assert job_name == expected_job_name
+
+
+@pytest.mark.parametrize(
+    "key,expected_key",
+    [
+        ('jobs.container-test-job.CKV_GHA_3[7:23]', "jobs.container-test-job"),
+        ('jobs.*.steps[].jobs.*.steps[].CKV_GHA_3[18:23]', "jobs.container-test-job.steps"),
+    ],
+)
+def test_modify_gha_key(key, expected_key, definition):
+    check = SuspectCurlInScript()
+    check.bc_id = 'GITHUB_ACTION_3'
+
+    new_key = Registry.modify_gha_key(key, check, definition)
+
+    assert new_key == expected_key
